+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [commands] Add ignore case for map based parser
 - [common] Add Rotation and Transform classes
 - [docs] Add docs talking about region box, transform, region and rotation.
+- [spigot-commands] `MaterialParser` using namespaced key for tab competition
 
 ## v0.0.29
 - [commands] Improve location parser